chain, graph: Move BlockStream from chain::ethereum to graph
