+++ chain/ethereum/src/lib.rs
@@ -7,7 +7,7 @@ mod transport;
 
 use graph::prelude::{BlockPtr, EthereumBlockWithTriggers};
 
-pub use self::block_stream::{BlockStream, BlockStreamBuilder};
+pub use self::block_stream::BlockStreamBuilder;
 pub use self::ethereum_adapter::{triggers_in_block, EthereumAdapter};
 pub use self::transport::{EventLoopHandle, Transport};
 
@@ -18,7 +18,6 @@ pub use crate::adapter::{
     EthereumAdapter as EthereumAdapterTrait, EthereumContractCall, EthereumContractCallError,
     MockEthereumAdapter, ProviderEthRpcMetrics, SubgraphEthRpcMetrics, TriggerFilter,
 };
-pub use crate::block_stream::BlockStreamMetrics;
 pub use crate::chain::Chain;
 pub use crate::network::EthereumNetworks;
